rewrite json testing stuff
