+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/keyvalue/api/watch/LockWatchEventCacheIntegrationTest.java
@@ -48,7 +48,6 @@ import com.palantir.lock.watch.ImmutableTransactionUpdate;
 import com.palantir.lock.watch.LockEvent;
 import com.palantir.lock.watch.LockWatchCreatedEvent;
 import com.palantir.lock.watch.LockWatchEvent;
-import com.palantir.lock.watch.LockWatchEventCache;
 import com.palantir.lock.watch.LockWatchReferences;
 import com.palantir.lock.watch.LockWatchStateUpdate;
 import com.palantir.lock.watch.TransactionUpdate;
@@ -58,8 +57,11 @@ import com.palantir.logsafe.exceptions.SafeIllegalArgumentException;
 
 public class LockWatchEventCacheIntegrationTest {
     private static final String TABLE = "table";
+    // Serializes as "dGFibGUAAQ=="
     private static final LockDescriptor DESCRIPTOR = AtlasRowLockDescriptor.of(TABLE, new byte[] {1});
+    // Serializes as "dGFibGUAAg=="
     private static final LockDescriptor DESCRIPTOR_2 = AtlasRowLockDescriptor.of(TABLE, new byte[] {2});
+    // Serializes as "dGFibGUAAw=="
     private static final LockDescriptor DESCRIPTOR_3 = AtlasRowLockDescriptor.of(TABLE, new byte[] {3});
     private static final LockWatchReferences.LockWatchReference REFERENCE = LockWatchReferences.entireTable("table");
     private static final UUID COMMIT_UUID = UUID.fromString("203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59");
@@ -72,9 +74,11 @@ public class LockWatchEventCacheIntegrationTest {
             LockEvent.builder(ImmutableSet.of(DESCRIPTOR_3), COMMIT_TOKEN).build(6L);
     private static final UUID EVENT2_UUID = UUID.fromString("888fcd7a-b3d7-4d2a-9d2c-3d61cde1ba44");
     private static final LockWatchEvent LOCK_EVENT_2 =
-            LockEvent.builder(ImmutableSet.of(DESCRIPTOR), LockToken.of(EVENT2_UUID)).build(10L);
+            LockEvent.builder(ImmutableSet.of(DESCRIPTOR), LockToken.of(EVENT2_UUID)).build(7L);
 
     private static final UUID LEADER = UUID.fromString("470c855e-f77b-44df-b56a-14d3df085dbc");
+    public static final LockWatchStateUpdate.Success SUCCESS_2 = LockWatchStateUpdate.success(LEADER, 7L,
+            ImmutableList.of(LOCK_EVENT_2));
     private static final LockWatchStateUpdate SNAPSHOT =
             LockWatchStateUpdate.snapshot(LEADER, 3L, ImmutableSet.of(DESCRIPTOR_2), ImmutableSet.of());
     private static final LockWatchStateUpdate SUCCESS =
@@ -97,7 +101,7 @@ public class LockWatchEventCacheIntegrationTest {
 
     }
 
-    private LockWatchEventCache eventCache;
+    private LockWatchEventCacheImpl eventCache;
     private int part;
 
     @Rule
@@ -105,7 +109,7 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Before
     public void before() {
-        eventCache = new LockWatchEventCacheImpl(LockWatchEventLogImpl.create());
+        eventCache = new LockWatchEventCacheImpl(LockWatchEventLog.create());
         part = 1;
     }
 
@@ -116,11 +120,12 @@ public class LockWatchEventCacheIntegrationTest {
                 .registerModule(new GuavaModule());
         try {
             Path path = Paths.get(BASE + name.getMethodName() + "/event-cache-" + part + ".json");
+            LockWatchEventCacheState eventCacheState = eventCache.getStateForTesting();
 
             if (MODE.isDev()) {
-                mapper.writeValue(path.toFile(), eventCache);
+                mapper.writeValue(path.toFile(), eventCacheState);
             } else {
-                String ourJson = mapper.writeValueAsString(eventCache);
+                String ourJson = mapper.writeValueAsString(eventCacheState);
                 String theirJson = new String(Files.readAllBytes(path), Charset.defaultCharset());
                 assertThat(ourJson).isEqualTo(theirJson);
             }
@@ -132,17 +137,65 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void processStartTimestampUpdateOnMultipleBatches() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         verifyStage();
 
         Set<Long> secondTimestamps = ImmutableSet.of(5L, 123L);
         eventCache.processStartTransactionsUpdate(secondTimestamps, SUCCESS);
         verifyStage();
+
+        Set<Long> thirdTimestamps = ImmutableSet.of(6677L, 8888L);
+        eventCache.processStartTransactionsUpdate(thirdTimestamps, SUCCESS_2);
+        verifyStage();
     }
 
     @Test
+    public void sameUpdateTwiceButDifferentTimestamps() {
+        setupInitialState();
+        Set<Long> secondTimestamps = ImmutableSet.of(11L, 12L);
+        Set<Long> thirdTimestamps = ImmutableSet.of(91L, 92L, 93L);
+
+        eventCache.processStartTransactionsUpdate(secondTimestamps, SUCCESS);
+        verifyStage();
+
+        eventCache.processStartTransactionsUpdate(thirdTimestamps, SUCCESS);
+        verifyStage();
+    }
+
+    @Test
+    public void smallerUpdateAfterLargeUpdateDoesNotAffectCache() {
+        setupInitialState();
+        Set<Long> secondTimestamps = ImmutableSet.of(11L, 12L);
+        Set<Long> thirdTimestamps = ImmutableSet.of(91L, 92L, 93L);
+
+        eventCache.processStartTransactionsUpdate(secondTimestamps, LockWatchStateUpdate.success(LEADER, 7L,
+                ImmutableList.of(WATCH_EVENT, UNLOCK_EVENT, LOCK_EVENT, LOCK_EVENT_2)));
+        verifyStage();
+
+        eventCache.processStartTransactionsUpdate(thirdTimestamps, LockWatchStateUpdate.success(LEADER, 5L,
+                ImmutableList.of(WATCH_EVENT, UNLOCK_EVENT)));
+        verifyStage();
+    }
+
+    @Test
+    public void largerUpdateAfterSmallUpdateOnlyPicksUpNewEvents() {
+        setupInitialState();
+        Set<Long> secondTimestamps = ImmutableSet.of(11L, 12L);
+        Set<Long> thirdTimestamps = ImmutableSet.of(91L, 92L, 93L);
+
+        eventCache.processStartTransactionsUpdate(secondTimestamps, LockWatchStateUpdate.success(LEADER, 5L,
+                ImmutableList.of(WATCH_EVENT, UNLOCK_EVENT)));
+        verifyStage();
+
+        eventCache.processStartTransactionsUpdate(thirdTimestamps, LockWatchStateUpdate.success(LEADER, 7L,
+                ImmutableList.of(WATCH_EVENT, UNLOCK_EVENT, LOCK_EVENT, LOCK_EVENT_2)));
+        verifyStage();
+    }
+
+
+    @Test
     public void getCommitUpdateDoesNotContainCommitLocks() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processGetCommitTimestampsUpdate(COMMIT_UPDATE, SUCCESS);
         verifyStage();
 
@@ -153,7 +206,7 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void cacheClearedOnSnapshotUpdate() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, SUCCESS);
         verifyStage();
 
@@ -166,19 +219,18 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void getEventsForTransactionsReturnsSnapshotWithOldEvents() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processGetCommitTimestampsUpdate(COMMIT_UPDATE, SUCCESS);
         eventCache.removeTransactionStateFromCache(START_TS);
         verifyStage();
 
-        LockWatchStateUpdate success2 = LockWatchStateUpdate.success(LEADER, 10L, ImmutableList.of(LOCK_EVENT_2));
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, success2);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, SUCCESS_2);
         verifyStage();
 
         TransactionsLockWatchUpdate results = eventCache.getUpdateForTransactions(TIMESTAMPS_2, Optional.empty());
         assertThat(results.clearCache()).isTrue();
         assertThat(results.startTsToSequence()).containsExactlyInAnyOrderEntriesOf(
-                ImmutableMap.of(16L, IdentifiedVersion.of(LEADER, 10L)));
+                ImmutableMap.of(16L, IdentifiedVersion.of(LEADER, 7L)));
         assertThat(results.events()).containsExactly(
                 LockWatchCreatedEvent.builder(ImmutableSet.of(REFERENCE),
                         ImmutableSet.of(DESCRIPTOR, DESCRIPTOR_3)).build(6L),
@@ -187,14 +239,14 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void failedUpdateClearsAllCaches() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, LockWatchStateUpdate.failed(LEADER));
         verifyStage();
     }
 
     @Test
     public void leaderChangeClearsCaches() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
                 LockWatchStateUpdate.success(EVENT2_UUID, 4L, ImmutableList.of()));
         verifyStage();
@@ -202,7 +254,7 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void removingEntriesRetentionsEventsInLog() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        setupInitialState();
         eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, SUCCESS);
         verifyStage();
 
@@ -212,32 +264,24 @@ public class LockWatchEventCacheIntegrationTest {
 
     @Test
     public void nonContiguousEventsThrows() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
-                LockWatchStateUpdate.success(LEADER, 10L, ImmutableList.of(WATCH_EVENT, LOCK_EVENT, LOCK_EVENT_2)));
-        eventCache.processStartTransactionsUpdate(ImmutableSet.of(23L, 55L),
-                LockWatchStateUpdate.success(LEADER, 20L,
-                        ImmutableList.of(UnlockEvent.builder(ImmutableSet.of()).build(20L))));
-        verifyStage();
-        assertThatThrownBy(() -> eventCache.removeTransactionStateFromCache(START_TS))
+        setupInitialState();
+        assertThatThrownBy(() -> eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
+                LockWatchStateUpdate.success(LEADER, 7L, ImmutableList.of(WATCH_EVENT, LOCK_EVENT, LOCK_EVENT_2))))
                 .isExactlyInstanceOf(SafeIllegalArgumentException.class)
                 .hasMessage("Events form a non-contiguous sequence");
     }
 
     @Test
-    public void snapshotMissedEventThrows() {
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
-        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
-                LockWatchStateUpdate.success(LEADER, 5L, ImmutableList.of(UNLOCK_EVENT)));
-        eventCache.processStartTransactionsUpdate(ImmutableSet.of(23L, 55L),
-                LockWatchStateUpdate.success(LEADER, 6L, ImmutableList.of(LOCK_EVENT)));
-        verifyStage();
-
-        eventCache.removeTransactionStateFromCache(START_TS);
-        verifyStage();
-        assertThatThrownBy(() -> eventCache.removeTransactionStateFromCache(16L))
+    public void missedEventThrows() {
+        setupInitialState();
+        assertThatThrownBy(() -> eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
+                LockWatchStateUpdate.success(LEADER, 5L, ImmutableList.of(UNLOCK_EVENT))))
                 .isExactlyInstanceOf(SafeIllegalArgumentException.class)
                 .hasMessage("Events missing between last snapshot and this batch of events");
+    }
+
+    private void setupInitialState() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
     }
 
     private static final class CommitUpdateVisitor implements CommitUpdate.Visitor<Set<LockDescriptor>> {
