+++ atlasdb-impl-shared/src/test/resources/lockwatch-event-cache-output/leaderChangeClearsCaches/event-cache-1.json
@@ -1,16 +1,16 @@
 {
-  "eventLog" : {
-    "snapshot" : {
+  "logState" : {
+    "snapshotState" : {
       "watches" : [ ],
       "locked" : [ ],
       "snapshotVersion" : null
     },
-    "eventStore" : {
+    "eventStoreState" : {
       "eventMap" : { }
     },
     "latestVersion" : null
   },
-  "timestampStateStore" : {
+  "timestampStoreState" : {
     "timestampMap" : { },
     "aliveVersions" : { }
   }
