+++ atlasdb-impl-shared/src/test/resources/lockwatch-event-cache-output/nonContiguousEventsThrows/event-cache-1.json
@@ -1,6 +1,6 @@
 {
-  "eventLog" : {
-    "snapshot" : {
+  "logState" : {
+    "snapshotState" : {
       "watches" : [ ],
       "locked" : [ {
         "bytes" : "dGFibGUAAg=="
@@ -10,7 +10,7 @@
         "version" : 3
       }
     },
-    "eventStore" : {
+    "eventStoreState" : {
       "eventMap" : {
         "4" : {
           "type" : "created",
@@ -33,9 +33,9 @@
             "requestId" : "203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59"
           }
         },
-        "10" : {
+        "7" : {
           "type" : "lock",
-          "sequence" : 10,
+          "sequence" : 7,
           "lockDescriptors" : [ {
             "bytes" : "dGFibGUAAQ=="
           } ],
@@ -55,12 +55,12 @@
       "version" : 20
     }
   },
-  "timestampStateStore" : {
+  "timestampStoreState" : {
     "timestampMap" : {
       "16" : {
         "version" : {
           "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
-          "version" : 10
+          "version" : 7
         },
         "commitInfo" : null
       },
@@ -88,7 +88,7 @@
     },
     "aliveVersions" : {
       "3" : [ 1 ],
-      "10" : [ 16 ],
+      "7" : [ 16 ],
       "20" : [ 23, 55 ]
     }
   }
