+++ atlasdb-impl-shared/src/test/resources/lockwatch-event-cache-output/processStartTimestampUpdateOnMultipleBatches/event-cache-3.json
@@ -39,12 +39,22 @@
           "lockToken" : {
             "requestId" : "203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59"
           }
+        },
+        "7" : {
+          "type" : "lock",
+          "sequence" : 7,
+          "lockDescriptors" : [ {
+            "bytes" : "dGFibGUAAQ=="
+          } ],
+          "lockToken" : {
+            "requestId" : "888fcd7a-b3d7-4d2a-9d2c-3d61cde1ba44"
+          }
         }
       }
     },
     "latestVersion" : {
       "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
-      "version" : 6
+      "version" : 7
     }
   },
   "timestampStoreState" : {
@@ -54,19 +64,41 @@
           "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
           "version" : 3
         },
-        "commitInfo" : {
-          "commitLockToken" : {
-            "requestId" : "203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59"
-          },
-          "commitVersion" : {
-            "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
-            "version" : 6
-          }
-        }
+        "commitInfo" : null
+      },
+      "5" : {
+        "version" : {
+          "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
+          "version" : 6
+        },
+        "commitInfo" : null
+      },
+      "6677" : {
+        "version" : {
+          "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
+          "version" : 7
+        },
+        "commitInfo" : null
+      },
+      "8888" : {
+        "version" : {
+          "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
+          "version" : 7
+        },
+        "commitInfo" : null
+      },
+      "123" : {
+        "version" : {
+          "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
+          "version" : 6
+        },
+        "commitInfo" : null
       }
     },
     "aliveVersions" : {
-      "3" : [ 1 ]
+      "3" : [ 1 ],
+      "6" : [ 5, 123 ],
+      "7" : [ 6677, 8888 ]
     }
   }
 }
